refactor: improve code layering of ozhera-webhook module
